+++ engine/engine/evaluation.h
@@ -4,4 +4,4 @@
 bool IsThreefoldRepetition(Game& game);
 unsigned char GetGameState(Game& game); //TODO: add threefold repetition and insufficient material
 int EvaluateMaterial(Game& game);
-float EvaluatePosition(Game& game);
+int EvaluatePosition(Game& game);
